requirements.txt
datasets==2.14.4
wandb==0.15.9
evaluate==0.4.0
hydra-core==1.3.2
matplotlib==3.7.1
numpy==1.23.5
omegaconf==2.3.0
pandas==2.0.1
PyYAML==6.0
setuptools==41.2.0
tensorboard==2.12.2
torch==2.0.0
tqdm==4.65.0
transformers==4.28.1
scikit-learn==1.3.0
seaborn==0.12.2
future